+++ base/authority/src/main/java/com/kewen/framework/base/authority/service/impl/SysUserDeptServiceImpl.java
@@ -45,7 +45,7 @@ public class SysUserDeptServiceImpl extends ServiceImpl<SysUserDeptMapper, SysUs
     }
 
     @Override
-    public UserDept getUserDept(Long userId) {
+    public UserDept getUserDept(Integer userId) {
         List<DeptPrimary> depts = baseMapper.listUserDept(userId);
 
         if (depts==null){
